[Feat] Fold Voice into the Chat tab as a mode, ChatGPT-style
Voice and Chat were two tabs holding the same conversation, and Chat already
had a mic button, so the app offered three ways to talk in two places. Now
there is one Chat tab: dictation types into the composer, and a second button
opens a full-screen spoken conversation over it.

- VoiceMode replaces ContinuousVoiceWorkspace. Same WebRTC realtime session,
  restyled onto the chat palette so entering voice reads as the same thread
  rather than a different product. Phase is legible from the orb's animation,
  not just its label.
- Voice turns join the chat thread on exit, so a spoken conversation leaves
  something behind instead of vanishing with the overlay. Half-streamed turns
  are dropped so no fragment lands in the thread.
- /voice is a 308 in next.config.ts rather than a redirect() inside a page.
  Redirects are checked before the filesystem, so bookmarks and crawlers get
  a real HTTP redirect; the static page version answered 200 and only moved
  the reader once JavaScript ran.
- Removed the Voice nav item and the home page's "Open Voice Workspace"
  button, and dropped 32 now-orphaned .voice-workspace/-stage/-transcript/
  -controls rules plus the unused voice-level keyframes from globals.css.
  .voice-orb and .voice-input stay — dictation still uses them.

KNOWN GAP, not closed here: the realtime session streams from the browser
straight to OpenAI and never passes through /api/chat, so the deterministic
safety floor and the model_runs audit row that text chat gained in #63 do not
apply to spoken conversation. Carrying the transcript into the thread makes
the gap visible rather than hidden, but closing it needs the turns posted
server-side. Worth doing before voice reaches students.
