--- sentra/frontend/src/app/voice/page.tsx
@@ -1,5 +1,0 @@
-import { ContinuousVoiceWorkspace } from "@/components/ContinuousVoiceWorkspace";
-
-export default function VoicePage() {
-  return <ContinuousVoiceWorkspace />;
-}
