+++ sentra/frontend/src/components/VoiceMode.module.css
@@ -1,0 +1,211 @@
+/* Full-screen voice mode, opened from the chat composer. Uses the chat
+   palette so entering voice reads as the same conversation, not a new place. */
+
+.overlay {
+  position: fixed;
+  inset: 0;
+  z-index: 60;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 24px 20px calc(28px + env(safe-area-inset-bottom));
+  background: linear-gradient(180deg, hsla(var(--hue), 60%, 97%, 0.97) 0%, var(--blue-mist) 100%);
+  backdrop-filter: blur(24px) saturate(1.3);
+  -webkit-backdrop-filter: blur(24px) saturate(1.3);
+  animation: fade 0.28s ease;
+}
+
+@keyframes fade {
+  from { opacity: 0; }
+  to { opacity: 1; }
+}
+
+.close {
+  position: absolute;
+  top: calc(18px + env(safe-area-inset-top));
+  right: 18px;
+  display: grid;
+  place-items: center;
+  width: 40px;
+  height: 40px;
+  border: none;
+  border-radius: 999px;
+  background: hsla(0, 0%, 100%, 0.7);
+  color: var(--blue-ink);
+  cursor: pointer;
+}
+
+.close:hover {
+  background: #ffffff;
+}
+
+/* Stage --------------------------------------------------------------- */
+
+.stage {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 14px;
+  margin-top: clamp(24px, 8vh, 72px);
+  text-align: center;
+}
+
+.orb {
+  display: grid;
+  place-items: center;
+  width: 168px;
+  height: 168px;
+  border-radius: 999px;
+  background: hsla(0, 0%, 100%, 0.66);
+  box-shadow: 0 10px 44px hsla(var(--hue), 60%, 50%, 0.16);
+}
+
+.orbFlower {
+  width: 104px;
+  height: auto;
+}
+
+/* One ring per phase rather than a generic pulse, so the state is legible
+   without reading the label. */
+.connecting { animation: breathe 1.6s ease-in-out infinite; }
+.listening  { box-shadow: 0 0 0 0 hsla(var(--hue), 74%, 62%, 0.5); animation: ripple 2s ease-out infinite; }
+.thinking   { animation: breathe 1.1s ease-in-out infinite; }
+.speaking   { animation: speak 0.9s ease-in-out infinite; }
+.interrupted, .error { box-shadow: 0 10px 44px hsla(8, 60%, 50%, 0.2); }
+
+@keyframes breathe {
+  0%, 100% { transform: scale(1); }
+  50%      { transform: scale(1.045); }
+}
+
+@keyframes ripple {
+  0%   { box-shadow: 0 0 0 0 hsla(var(--hue), 74%, 62%, 0.45); }
+  100% { box-shadow: 0 0 0 42px hsla(var(--hue), 74%, 62%, 0); }
+}
+
+@keyframes speak {
+  0%, 100% { transform: scale(1); }
+  30%      { transform: scale(1.07); }
+  60%      { transform: scale(1.02); }
+}
+
+.phase {
+  font-size: 1.15rem;
+  font-weight: 600;
+  color: var(--blue-ink);
+}
+
+.hint {
+  max-width: 30ch;
+  margin: 0;
+  font-size: 0.9rem;
+  line-height: 1.6;
+  color: hsla(var(--hue), 40%, 32%, 0.85);
+}
+
+.error {
+  max-width: 34ch;
+  font-size: 0.9rem;
+  color: hsl(8, 62%, 42%);
+}
+
+/* Transcript ----------------------------------------------------------- */
+
+.transcript {
+  flex: 1;
+  width: 100%;
+  max-width: 640px;
+  margin-top: 26px;
+  overflow-y: auto;
+  display: flex;
+  flex-direction: column;
+  gap: 14px;
+  padding: 0 4px 12px;
+  -webkit-mask-image: linear-gradient(180deg, transparent 0, #000 22px, #000 calc(100% - 12px), transparent 100%);
+  mask-image: linear-gradient(180deg, transparent 0, #000 22px, #000 calc(100% - 12px), transparent 100%);
+}
+
+.line {
+  display: flex;
+  flex-direction: column;
+  gap: 3px;
+}
+
+.line p {
+  margin: 0;
+  font-size: 0.98rem;
+  line-height: 1.65;
+  color: var(--blue-ink);
+}
+
+.lineUser {
+  align-items: flex-end;
+  text-align: right;
+}
+
+.who {
+  font-size: 0.72rem;
+  letter-spacing: 0.08em;
+  text-transform: uppercase;
+  color: hsla(var(--hue), 30%, 45%, 0.8);
+}
+
+.empty {
+  margin: 0;
+}
+
+/* Controls ------------------------------------------------------------- */
+
+.controls {
+  display: flex;
+  align-items: center;
+  gap: 22px;
+  margin-top: 18px;
+}
+
+.control {
+  display: grid;
+  place-items: center;
+  width: 52px;
+  height: 52px;
+  border: 1px solid hsla(0, 0%, 100%, 0.7);
+  border-radius: 999px;
+  background: hsla(0, 0%, 100%, 0.6);
+  color: var(--blue-ink);
+  cursor: pointer;
+  transition: background 0.2s ease, opacity 0.2s ease;
+}
+
+.control:hover:not(:disabled) { background: #ffffff; }
+.control:disabled { opacity: 0.4; cursor: default; }
+
+.primary {
+  display: grid;
+  place-items: center;
+  width: 76px;
+  height: 76px;
+  border: none;
+  border-radius: 999px;
+  background: var(--blue-base);
+  color: #ffffff;
+  cursor: pointer;
+  box-shadow: 0 8px 26px hsla(var(--hue), 60%, 46%, 0.3);
+  transition: background 0.2s ease, transform 0.2s ease;
+}
+
+.primary:hover { background: var(--blue-hover); }
+.primary:active { transform: scale(0.96); }
+.primaryActive { background: hsl(8, 62%, 56%); box-shadow: 0 8px 26px hsla(8, 62%, 46%, 0.3); }
+.primaryActive:hover { background: hsl(8, 62%, 48%); }
+
+.disclaimer {
+  margin: 14px 0 0;
+  font-size: 0.74rem;
+  color: hsla(var(--hue), 26%, 46%, 0.9);
+  text-align: center;
+}
+
+@media (prefers-reduced-motion: reduce) {
+  .overlay { animation: none; }
+  .connecting, .listening, .thinking, .speaking { animation: none; }
+}
